+++ src/pages/Member/member.scss
@@ -10,7 +10,7 @@ $shadow:#E7E7E7;
 
 @import url('https://fonts.googleapis.com/css2?family=Noto+Sans+TC:wght@300;400;500&display=swap');
 
-* {
+.member {
     .font-s {
         font-size: 1rem;
     }
